添加申请提交检查与园区及床位生成逻辑 修正building_id和building_num的数据类型 移除CodeGenerator类并加入park的添加

+++ src/main/java/com/scu/accommodationmanagement/model/po/Park.java
@@ -46,7 +46,7 @@ public class Park implements Serializable {
      * 园区内楼栋数量
      */
     @TableField("building_num")
-    private Long buildingNum;
+    private Integer buildingNum;
 
 
 }
